Move benchmarking code to a library.
